src/requirements.txt
PyYAML==6.0.1
torch==2.2.1
